+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Changelog
 =========
-## Unreleased
+## 1.0.0-beta.4 - 2018-04-29
 ### Added
 - View Url property to the SObject field which supports direct linking to Salesforce Object
 - List Url property to the SObject field which supports direct linking to Salesforce Object List